changed discount data type from int to float

+++ ITSC3155-Assignments-main 2/FinalProject/api/models/promos.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer, String, DECIMAL, DATETIME
+from sqlalchemy import Column, ForeignKey, Integer, String, Float, DECIMAL, DATETIME
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from ..dependencies.database import Base
@@ -10,6 +10,6 @@ class Promo(Base):
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     code = Column(String(100))
     menu_id = Column(Integer, ForeignKey("menu.id"))
-    discount = Column(Integer)
+    discount = Column(Float)
 
     menu_item = relationship("Item", back_populates="menu")
